Port Netcdf4 jna reader. Pass 1, writer not working yet.
