+++ cdm/core/src/main/java/ucar/nc2/iosp/AbstractIOServiceProvider.java
@@ -54,7 +54,7 @@ public abstract class AbstractIOServiceProvider implements IOServiceProvider {
   // TODO this is misused, probably should be in a constructor?
   public void open(RandomAccessFile raf, NetcdfFile ncfile, CancelTask cancelTask) throws IOException {
     this.raf = raf;
-    this.location = (raf != null) ? raf.getLocation() : null;
+    this.location = (raf != null) ? raf.getLocation() : ncfile != null ? ncfile.getLocation() : null;
     this.ncfile = ncfile; // LOOK do we need this? why ?
   }
