+++ cdm/core/src/main/java/ucar/nc2/iosp/IospHelper.java
@@ -258,8 +258,12 @@ public class IospHelper {
    */
   public static Object readDataFill(LayoutBB layout, DataType dataType, Object fillValue) {
     long size = layout.getTotalNelems();
-    if (dataType == DataType.STRUCTURE)
+    if (dataType == DataType.STRUCTURE) {
       size *= layout.getElemSize();
+    }
+    if (size >= Integer.MAX_VALUE) {
+      throw new RuntimeException("Read request too large");
+    }
     Object arr = (fillValue == null) ? makePrimitiveArray((int) size, dataType)
         : makePrimitiveArray((int) size, dataType, fillValue);
     return readData(layout, dataType, arr);
