+++ netcdf4/src/main/java/ucar/nc2/ffi/netcdf/Nc4wrapper.java
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) 1998-2020 John Caron and University Corporation for Atmospheric Research/Unidata
- *  See LICENSE for license information.
+ * See LICENSE for license information.
  */
 
 package ucar.nc2.ffi.netcdf;
