+++ netcdf4/src/test/java/ucar/nc2/jni/netcdf/TestCDF5Reading.java
@@ -5,36 +5,29 @@
 
 package ucar.nc2.jni.netcdf;
 
-import org.junit.Assert;
+import static com.google.common.truth.Truth.assertThat;
+import static com.google.common.truth.Truth.assertWithMessage;
+
 import org.junit.Assume;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import ucar.ma2.Array;
 import ucar.ma2.ArrayFloat;
 import ucar.ma2.InvalidRangeException;
 import ucar.ma2.MAMath;
 import ucar.nc2.NetcdfFile;
-import ucar.nc2.NetcdfFileSubclass;
 import ucar.nc2.Variable;
 import ucar.nc2.ffi.netcdf.NetcdfClibrary;
 import ucar.nc2.iosp.NetcdfFileFormat;
-import ucar.nc2.write.Ncdump;
 import ucar.unidata.io.RandomAccessFile;
 import ucar.unidata.util.test.TestDir;
-import ucar.unidata.util.test.UnitTestCommon;
 import ucar.unidata.util.test.category.NeedsContentRoot;
 import java.io.IOException;
-import java.lang.invoke.MethodHandles;
 
-/**
- * Test Reading of CDF-5 files using JNI netcdf-4 iosp
- */
+/** Test Reading of CDF-5 files using JNI netcdf-4 iosp */
 @Category(NeedsContentRoot.class)
-public class TestCDF5Reading extends UnitTestCommon {
-  private static final Logger logger = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
+public class TestCDF5Reading {
   private static final ArrayFloat.D1 BASELINE;
 
   static {
@@ -54,22 +47,15 @@ public class TestCDF5Reading extends UnitTestCommon {
 
   @Test
   public void testReadSubsection() throws IOException, InvalidRangeException {
-    String location = canonjoin(TestDir.cdmTestDataDir, "thredds/public/testdata/nc_test_cdf5.nc");
+    String location = TestDir.cdmTestDataDir + "thredds/public/testdata/nc_test_cdf5.nc";
     try (RandomAccessFile raf = RandomAccessFile.acquire(location)) {
       // Verify that this is a netcdf-5 file
       NetcdfFileFormat format = NetcdfFileFormat.findNetcdfFormatType(raf);
-      Assert.assertTrue("Fail: file format is not CDF-5", format == NetcdfFileFormat.NETCDF3_64BIT_DATA);
+      assertWithMessage("Fail: file format is not CDF-5").that(format).isEqualTo(NetcdfFileFormat.NETCDF3_64BIT_DATA);
     }
-    try (NetcdfFile jni = openJni(location)) {
-      jni.setLocation(location + " (jni)");
+    try (NetcdfFile jni = TestNc4reader.openJni(location)) {
       Array data = read(jni, "f4", "0:2");
-      if (prop_visual) {
-        String dump = Ncdump.printArray(data);
-        logger.debug(dump);
-        String testresult = dump.replace('r', ' ').replace('\n', ' ').trim();
-        visual("CDF Read", testresult);
-      }
-      Assert.assertTrue("***Fail: data mismatch", MAMath.nearlyEquals(data, BASELINE));
+      assertThat(MAMath.nearlyEquals(data, BASELINE)).isTrue();
       System.err.println("***Pass");
     }
   }
@@ -78,14 +64,6 @@ public class TestCDF5Reading extends UnitTestCommon {
     Variable v = ncfile.findVariable(vname);
     assert v != null;
     return v.read(section);
-  }
-
-  private NetcdfFile openJni(String location) throws IOException {
-    Nc4Iosp iosp = new Nc4Iosp();
-    NetcdfFile ncfile = new NetcdfFileSubclass(iosp, location);
-    RandomAccessFile raf = new RandomAccessFile(location, "r");
-    iosp.open(raf, ncfile, null);
-    return ncfile;
   }
 
 }
