+++ uicdm/src/main/java/ucar/nc2/ui/ToolsUI.java
@@ -14,7 +14,6 @@ import ucar.httpservices.HTTPSession;
 import ucar.nc2.*;
 import ucar.nc2.constants.FeatureType;
 import ucar.nc2.dataset.*;
-// import ucar.nc2.dods.DODSNetcdfFile;
 import ucar.nc2.dt.GridDataset;
 import ucar.nc2.dt.RadialDatasetSweep;
 import ucar.nc2.ft.point.PointDatasetImpl;
@@ -24,7 +23,6 @@ import ucar.nc2.grib.collection.GribCdmIndex;
 import ucar.nc2.internal.iosp.hdf5.H5iospNew;
 import ucar.nc2.internal.ncml.Aggregation;
 import ucar.nc2.internal.ncml.NcmlReader;
-// import ucar.nc2.jni.netcdf.Nc4Iosp;
 import ucar.nc2.stream.CdmRemote;
 import ucar.nc2.ui.dialog.DiskCache2Form;
 import ucar.nc2.ui.grib.*;
